padronização

+++ delete_sistema.php
@@ -5,11 +5,13 @@ if (!isset($_SESSION['username']) == true || $_SESSION["admin"] == 0) {
   header('location:index.php');
 }
 
-// Conecta-se ao banco de dados
 $conn = new mysqli('localhost', 'root', '', 'AHAHAHABORGES');
+if ($conn->connect_error) {
+  die("Conexão falhou: " . $conn->connect_error);
+}
+
 $id = isset($_GET['id']) ? $_GET['id'] : null;
 
-// Deleta o sistema
 $delete = "DELETE FROM sistemas WHERE id = '$id'";
 $result = $conn->query($delete);
